+++ example/lib/demo/demo_page.ctrl.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 
 import 'package:flutter/widgets.dart';
-import 'package:flutter_xml_layout_helpers/flutter_xml_layout_helpers.dart';
+import 'package:flutter_xml_layout_helpers/headers.dart';
 
 import '../locale_controller.dart';
 import 'demo_page.xml.dart';